gql: friend request mutation updates (#4024)
* reorganize resolvers

* update send friend request mutation
