+++ backend/native/backpack-api/src/routes/graphql/resolvers/index.ts
@@ -1,6 +1,2 @@
-export * from "./friendship";
-export * from "./mutations";
-export * from "./notification";
-export * from "./scalars";
-export * from "./user";
-export * from "./wallet";
+export * from "./mutation";
+export * from "./query";
