+++ backend/native/backpack-api/src/routes/graphql/resolvers/mutation/index.ts
@@ -1,0 +1,3 @@
+export * from "./authentication";
+export * from "./friendship";
+export * from "./wallets";
