+++ backend/native/backpack-api/src/routes/graphql/resolvers/query/index.ts
@@ -1,0 +1,5 @@
+export * from "./friendship";
+export * from "./notification";
+export * from "./scalars";
+export * from "./user";
+export * from "./wallet";
